postgres db restore script added
This commit adds a bash shell script to restore a Postgres database
SQL dump into an already created database. This script was written
to be used with the postgres_backup.sh script in this repository.

The scrpt will decompressed the compressed dump (made with pg_dump),
use the psql command to load the SQL dump, and finally recompress
the SQL dump backup file.

This commit also renames the Postgres backup script in this repository
from backup_postgres.sh to postgres_backup.sh.